% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity_hourly.R
\docType{data}
\name{electricity_hourly}
\alias{electricity_hourly}
\title{Hourly data from the UCI ElectricityLoadDiagrams20112014 Data Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10464248 rows and 3 columns.
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/ElectricityLoadDiagrams20112014}{UCI database}
}
\usage{
electricity_hourly
}
\description{
The dataset contains hourly mean KW per 15min of 370 clients.
We have processed the dataset so it only includes the active period of
each client. Ie, we take the period between the first non zero-entry and
the last non-zero entry.
}
\examples{
library(electricity)
str(electricity_hourly)
}
\keyword{datasets}
